+++ docs/CHANGES.rst
@@ -4,5 +4,7 @@ Changelog
 0.1-dev (unreleased)
 --------------------
 
+- Add send credit script to allow scripting of sending credit.
+  [davidjb]
 - Package created using templer
   [davidjb]